Package rename: digital_archive -> digiarch (#82)
* Change folder name

* Change imports in test files

* Change references in documentation files

* Change package name in main py files

* Change name in CircleCI config

* Change package name in pyproject.toml

* Fix reference in mypy.ini

* Remove python magic as dependency (this will also trigger new cache build)

* Bump version to 0.3.0 due to rename
