+++ .circleci/config.yml
@@ -41,19 +41,19 @@ jobs:
       - run:
           name: Check formatting using Black
           command: |
-            poetry run black --check digital_archive/
+            poetry run black --check digiarch/
             poetry run black --check tests/
       # Check linting using flake8 via Poetry
       - run:
           name: Check linting using flake8
           command: |
-            poetry run flake8 digital_archive/
+            poetry run flake8 digiarch/
             poetry run flake8 tests/
       # Check types using mypy via Poetry
       - run:
           name: Check types using mypy
           command: |
-            poetry run mypy digital_archive/
+            poetry run mypy digiarch/
             poetry run mypy tests/
   test:
     docker:
@@ -64,7 +64,7 @@ jobs:
       - run:
           name: Test and generate report using pytest with coverage
           command: |
-            poetry run pytest --cov=digital_archive --cov-report=xml
+            poetry run pytest --cov=digiarch --cov-report=xml
       # Upload coverage to codecov
       - codecov/upload:
           file: coverage.xml
@@ -73,7 +73,7 @@ jobs:
       - image: circleci/python:3.7
     steps:
       - init
-      # Build digital_archive with Poetry
+      # Build digiarch with Poetry
       - run:
           name: Build sdist and wheel with Poetry
           command: poetry build
