+++ digiarch/identify/reports.py
@@ -7,8 +7,8 @@
 # -----------------------------------------------------------------------------
 import os
 import pandas as pd
-from digital_archive.data import get_fileinfo_list
-from digital_archive.utils import click_utils
+from digiarch.data import get_fileinfo_list
+from digiarch.utils import click_utils
 from typing import List
 
 # -----------------------------------------------------------------------------
