+++ mypy.ini
@@ -4,11 +4,11 @@
 python_version = 3.7
 warn_return_any = True
 warn_unused_configs = True
-# Because we disallow untyped defs in digital_archive,
+# Because we disallow untyped defs in digiarch,
 # check_untyped_defs only has an impact in tests/
 check_untyped_defs = True
 
-[mypy-digital_archive.*]
+[mypy-digiarch.*]
 disallow_untyped_defs = True
 
 [mypy-pandas]
