+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
-name = "digital_archive"
-version = "0.2.1"
+name = "digiarch"
+version = "0.3.0"
 description = "Tools for the Digital Archive Project at Aarhus Stadsarkiv"
 authors = ["Nina Jensen <jnik@aarhus.dk>", "Aarhus Stadsarkiv <stadsarkiv@aarhus.dk>"]
 license = "GPL-3.0"
@@ -14,7 +14,6 @@ keywords = ["archiving", "preservation", "digital-archive"]
 python = "^3.7"
 pandas = "^0.25.1"
 click = "^7.0"
-python-magic = "^0.4.15"
 tqdm = "^4.36"
 dacite = "^1.0"
 
@@ -26,7 +25,7 @@ black = {version = "^18.3-alpha.0", allows-prereleases = true}
 flake8 = "^3.7"
 
 [tool.poetry.scripts]
-digiarch = "digital_archive.digiarch:cli"
+digiarch = "digiarch.digiarch:cli"
 
 [tool.black]
 line-length = 79
