+++ tests/test_data.py
@@ -1,6 +1,6 @@
 import json
 import pytest
-from digital_archive.data import FileInfo, encode_dataclass
+from digiarch.data import FileInfo, encode_dataclass
 
 
 @pytest.fixture
